refactor manual lint with a shared fct
