+++ crates/cairo-lint-core/src/fix.rs
@@ -446,14 +446,9 @@ impl Fixer {
 
     /// Rewrites a manual implementation of ok_or
     pub fn fix_manual_ok_or(&self, db: &dyn SyntaxGroup, node: SyntaxNode) -> String {
-        let expr_match = if let Expr::Match(expr_match) = Expr::from_syntax_node(db, node.clone()) {
-            expr_match
-        } else {
-            panic!("Expected a match expression");
-        };
+        let expr_match = ExprMatch::from_syntax_node(db, node.clone());
 
-        let val = expr_match.expr(db);
-        let option_var_name = match val {
+        let option_var_name = match expr_match.expr(db) {
             Expr::Path(path_expr) => path_expr.as_syntax_node().get_text_without_trivia(db),
             _ => panic!("Expected a variable or path in match expression"),
         };
@@ -463,24 +458,9 @@ impl Fixer {
             panic!("Expected exactly two arms in the match expression");
         }
 
-        let first_arm = &arms[0];
         let second_arm = &arms[1];
 
-        let first_pattern = &first_arm.patterns(db).elements(db)[0];
-        let second_pattern = &second_arm.patterns(db).elements(db)[0];
-
-        match first_pattern {
-            Pattern::Enum(enum_pattern) => {
-                let enum_name = enum_pattern.path(db).as_syntax_node().get_text_without_trivia(db);
-                match enum_name.as_str() {
-                    "Option::Some" => {}
-                    _ => panic!("Expected Option::Some enum pattern"),
-                }
-            }
-            _ => panic!("Expected an Option enum pattern"),
-        }
-
-        let none_arm_err = match second_pattern {
+        let none_arm_err = match &second_arm.patterns(db).elements(db)[0] {
             Pattern::Enum(enum_pattern) => {
                 let enum_name = enum_pattern.path(db).as_syntax_node().get_text_without_trivia(db);
                 match enum_name.as_str() {
