+++ crates/cairo-lint-core/src/lints/manual/mod.rs
@@ -1,1 +1,2 @@
+pub mod manual;
 pub mod manual_ok_or;
